pyproject.toml
[project]
name = "api"
description = "Some API"
authors = [
    {name = "Ananas", email = "ananas@ananas.moe"}
]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = []
classifiers = [
	'Operating System :: Microsoft :: Windows :: Windows 11',
    'Operating System :: Microsoft :: Windows :: Windows 10',
    'Operating System :: POSIX :: Linux',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
	'Programming Language :: Python :: 3.11'
]
dependencies = [
    "fastapi[all]",
    "uvicorn[standard]",
    "python-decouple",
    "Pillow"
]

dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "ruff",
    "build"
]

[project.urls]
GitHub = "https://github.com/r3tr0ananas/contribimg"
BugTracker = "https://github.com/r3tr0ananas/contribimg/issues"

[tool.setuptools.dynamic]
version = { attr = "api.__version__" }

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"